Move dataframe hypothesis strategy to its own module
